move utility functions over to utils
